pyproject.toml
[project]
name = "dreaming-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.5.8",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
]
